CHU-TronClassTool-CLI/requirements.txt
beautifulsoup4
pycryptodome
requests~=2.32.5
selenium~=4.37.0
pillow~=12.0.0
rich~=14.2.0
aiohttp~=3.13.2
numpy~=2.2.6
pyzbar~=0.1.9
PyQt6~=6.7.1
PyQt6_sip~=13.10.2
opencv-python~=4.12.0.88